Crate fedimint_wallet_client

Modules

api

backup 🔒

cli 🔒

client_db

config

deposit 🔒

Legacy, state-machine based peg-ins, replaced by pegin_monitor but retained for time being to ensure existing peg-ins complete.

endpoint_constants

envs

events

keys

pegin_monitor 🔒

Peg-in monitor: a task monitoring deposit addresses for peg-ins.

tweakable

txoproof

withdraw 🔒

Structs

BitcoinTransactionData

PegOut

PegOutFees

PegOutSignatureItem

Rbf

Allows a user to bump the fees of a PendingTransaction

SpendableUTXO

TxOutputSummary

A transaction output, either unspent or consumed

UnknownWalletInputVariantError

UnknownWalletOutputOutcomeVariantError

UnknownWalletOutputVariantError

WalletClientContext

WalletClientInit

WalletClientModule

WalletClientModuleData

The non-resource, just plain-data parts of WalletClientModule

WalletCommonInit

WalletInputV0

WalletInputV1

WalletModuleTypes

WalletOperationMeta

WalletOutputOutcomeV0

Contains the Bitcoin transaction id of the transaction created by the withdraw request

WalletSummary

Summary of the coins within the wallet.

WalletSummaryRequest 🔒

Enums

DepositStateV1

DepositStateV2

ProcessPegOutSigError

WalletClientStates

WalletConsensusItem

WalletInput

WalletInputError

WalletOperationMetaVariant

WalletOutput

WalletOutputError

WalletOutputOutcome

WalletOutputV0

WithdrawState

Constants

DEPRECATED_RBF_ERROR

FEERATE_MULTIPLIER_DEFAULT

To further mitigate the risk of a peg-out transaction getting stuck in the mempool, we multiply the feerate estimate returned from the backend by this value.

KIND

MODULE_CONSENSUS_VERSION

SAFE_DEPOSIT_MODULE_CONSENSUS_VERSION

Module consensus version that introduced support for processing Bitcoin transactions that exceed the ALEPH_BFT_UNIT_BYTE_LIMIT.

WALLET_TWEAK_CHILD_ID 🔒

Functions

get_next_peg_in_tweak_child_id 🔒

Returns the child index to derive the next peg-in tweak key from.

next_withdraw_state 🔒

poll_supports_safe_deposit_version 🔒

Polls the federation checking if the activated module consensus version supports safe deposits, saving the result in the db once it does.

proprietary_tweak_key

Type Aliases

PartialSig

PegInDescriptor