fedimint_wallet_client

Module pegin_monitor

Peg-in monitor: a task monitoring deposit addresses for peg-ins.

Structs

NextActions 🔒

A helper struct meant to combined data from all addresses/records into a single struct with all actionable data.

Enums

CheckOutcome 🔒

Outcome of checking a single deposit Bitcoin transaction output

Functions

check_and_claim_idx_pegins 🔒

check_for_deposits 🔒

check_idx_pegins 🔒

Query via btc rpc for a history of an address derived with tweak_idx and claim any peg-ins that are ready.

claim_peg_in 🔒

filter_onchain_deposit_outputs 🔒

run_peg_in_monitor 🔒

A deposit monitoring task