fedimint_wallet_client::pegin_monitor

Function run_peg_in_monitor

pub(crate) async fn run_peg_in_monitor(
    client_ctx: ClientContext<WalletClientModule>,
    db: Database,
    btc_rpc: DynBitcoindRpc,
    module_api: DynModuleApi,
    data: WalletClientModuleData,
    pegin_claimed_sender: Sender<()>,
    wakeup_receiver: Receiver<()>,
)

A deposit monitoring task

On the high level it maintains a list of derived addresses with some info like when is the next time to check for deposits on them.