fedimint_wallet_client

Struct WalletClientModule

pub struct WalletClientModule {
    pub(crate) data: WalletClientModuleData,
    pub(crate) db: Database,
    pub(crate) module_api: DynModuleApi,
    pub(crate) notifier: ModuleNotifier<WalletClientStates>,
    pub(crate) rpc: DynBitcoindRpc,
    pub(crate) client_ctx: ClientContext<Self>,
    pub(crate) pegin_monitor_wakeup_sender: Sender<()>,
    pub(crate) pegin_monitor_wakeup_receiver: Receiver<()>,
    pub(crate) pegin_claimed_sender: Sender<()>,
    pub(crate) pegin_claimed_receiver: Receiver<()>,
    pub(crate) task_group: TaskGroup,
    pub(crate) admin_auth: Option<ApiAuth>,
}

Fields
data: WalletClientModuleDatadb: Databasemodule_api: DynModuleApinotifier: ModuleNotifier<WalletClientStates>rpc: DynBitcoindRpcclient_ctx: ClientContext<Self>pegin_monitor_wakeup_sender: Sender<()>

Updated to wake up pegin monitor

pegin_monitor_wakeup_receiver: Receiver<()>pegin_claimed_sender: Sender<()>

Called every time a peg-in was claimed

pegin_claimed_receiver: Receiver<()>task_group: TaskGroupadmin_auth: Option<ApiAuth>
Implementations

impl WalletClientModule

pub(crate) fn cfg(&self) -> &WalletClientConfig

pub(crate) fn get_rpc_config(cfg: &WalletClientConfig) -> BitcoinRpcConfig

pub fn get_network(&self) -> Network

pub fn get_finality_delay(&self) -> u32

pub fn get_fee_consensus(&self) -> FeeConsensus

pub(crate) async fn allocate_deposit_address_inner( &self, dbtx: &mut DatabaseTransaction<'_>, ) -> (OperationId, Address, TweakIdx)

pub async fn get_withdraw_fees( &self, address: &Address, amount: Amount, ) -> Result<PegOutFees>

Fetches the fees that would need to be paid to make the withdraw request using Self::withdraw work right now.

Note that we do not receive a guarantee that these fees will be valid in the future, thus even the next second using these fees may fail. The caller should be prepared to retry with a new fee estimate.

pub async fn get_wallet_summary(&self) -> Result<WalletSummary>

Returns a summary of the wallet’s coins

pub async fn get_block_count_local(&self) -> Result<u32>

pub fn create_withdraw_output( &self, operation_id: OperationId, address: Address, amount: Amount, fees: PegOutFees, ) -> Result<ClientOutputBundle<WalletOutput, WalletClientStates>>

pub async fn peg_in(&self, req: PegInRequest) -> Result<PegInResponse>

pub async fn peg_out(&self, req: PegOutRequest) -> Result<PegOutResponse>

pub fn create_rbf_withdraw_output( &self, operation_id: OperationId, rbf: &Rbf, ) -> Result<ClientOutputBundle<WalletOutput, WalletClientStates>>

pub async fn btc_tx_has_no_size_limit(&self) -> FederationResult<bool>

pub async fn supports_safe_deposit(&self) -> bool

Returns true if the federation’s wallet module consensus version supports processing all deposits.

This method is safe to call offline, since it first attempts to read a key from the db that represents the client has previously been able to verify the wallet module consensus version. If the client has not verified the version, it must be online to fetch the latest wallet module consensus version.

pub async fn safe_allocate_deposit_address<M>( &self, extra_meta: M, ) -> Result<(OperationId, Address, TweakIdx)>
where M: Serialize + MaybeSend + MaybeSync,

Allocates a deposit address controlled by the federation, guaranteeing safe handling of all deposits, including on-chain transactions exceeding ALEPH_BFT_UNIT_BYTE_LIMIT.

Returns an error if the client has never been online to verify the federation’s wallet module consensus version supports processing all deposits.

pub async fn allocate_deposit_address_expert_only<M>( &self, extra_meta: M, ) -> Result<(OperationId, Address, TweakIdx)>
where M: Serialize + MaybeSend + MaybeSync,

Allocates a deposit address that is controlled by the federation.

This is an EXPERT ONLY method intended for power users such as Lightning gateways allocating liquidity, and we discourage exposing peg-in functionality to everyday users of a Fedimint wallet due to the following two limitations:

The transaction sending to this address needs to be smaller than 40KB in order for the peg-in to be claimable. If the transaction is too large, funds will be lost.

In the future, federations will also enforce a minimum peg-in amount to prevent accumulation of dust UTXOs. Peg-ins under this minimum cannot be claimed and funds will be lost.

Everyday users should rely on Lightning to move funds into the federation.

pub async fn subscribe_deposit( &self, operation_id: OperationId, ) -> Result<UpdateStreamOrOutcome<DepositStateV2>>

Returns a stream of updates about an ongoing deposit operation created with WalletClientModule::allocate_deposit_address_expert_only. Returns an error for old deposit operations created prior to the 0.4 release and not driven to completion yet. This should be rare enough that an indeterminate state is ok here.

pub async fn list_peg_in_tweak_idxes( &self, ) -> BTreeMap<TweakIdx, PegInTweakIndexData>

pub async fn find_tweak_idx_by_address( &self, address: Address<NetworkUnchecked>, ) -> Result<TweakIdx>

pub async fn find_tweak_idx_by_operation_id( &self, operation_id: OperationId, ) -> Result<TweakIdx>

pub async fn get_pegin_tweak_idx( &self, tweak_idx: TweakIdx, ) -> Result<PegInTweakIndexData>

pub async fn get_claimed_pegins( &self, dbtx: &mut DatabaseTransaction<'_>, tweak_idx: TweakIdx, ) -> Vec<(OutPoint, TransactionId, Vec<OutPoint>)>

pub async fn recheck_pegin_address_by_op_id( &self, operation_id: OperationId, ) -> Result<()>

Like Self::recheck_pegin_address but by operation_id

pub async fn recheck_pegin_address_by_address( &self, address: Address<NetworkUnchecked>, ) -> Result<()>

Schedule given address for immediate re-check for deposits

pub async fn recheck_pegin_address(&self, tweak_idx: TweakIdx) -> Result<()>

Schedule given address for immediate re-check for deposits

pub async fn await_num_deposits_by_operation_id( &self, operation_id: OperationId, num_deposits: usize, ) -> Result<()>

Await for num deposit by OperationId

pub async fn await_num_deposits_by_address( &self, address: Address<NetworkUnchecked>, num_deposits: usize, ) -> Result<()>

pub async fn await_num_deposits( &self, tweak_idx: TweakIdx, num_deposits: usize, ) -> Result<()>

pub async fn withdraw<M: Serialize + MaybeSend + MaybeSync>( &self, address: &Address, amount: Amount, fee: PegOutFees, extra_meta: M, ) -> Result<OperationId>

Attempt to withdraw a given amount of Bitcoin to a destination address. The caller has to supply the fee rate to be used which can be fetched using Self::get_withdraw_fees and should be acknowledged by the user since it can be unexpectedly high.

pub async fn rbf_withdraw<M: Serialize + MaybeSync + MaybeSend>( &self, rbf: Rbf, extra_meta: M, ) -> Result<OperationId>

👎Deprecated since 0.4.0: RBF withdrawals are rejected by the federation

Attempt to increase the fee of a onchain withdraw transaction using replace by fee (RBF). This can prevent transactions from getting stuck in the mempool

pub async fn subscribe_withdraw_updates( &self, operation_id: OperationId, ) -> Result<UpdateStreamOrOutcome<WithdrawState>>

pub(crate) fn admin_auth(&self) -> Result<ApiAuth>

pub async fn activate_consensus_version_voting(&self) -> Result<()>