Trait fedimint_wallet_client::WalletClientExt

pub trait WalletClientExt {
    fn get_deposit_address<'life0, 'async_trait>(
        &'life0 self,
        valid_until: SystemTime
    ) -> Pin<Box<dyn Future<Output = Result<(OperationId, Address)>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn subscribe_deposit_updates<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<'_, DepositState>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_withdraw_fee<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        amount: Amount
    ) -> Pin<Box<dyn Future<Output = Result<PegOutFees>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn withdraw<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        amount: Amount,
        fee: PegOutFees
    ) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn subscribe_withdraw_updates<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<'_, WithdrawState>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
}

Required Methods

fn get_deposit_address<'life0, 'async_trait>(
&'life0 self,
valid_until: SystemTime
) -> Pin<Box<dyn Future<Output = Result<(OperationId, Address)>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

fn subscribe_deposit_updates<'life0, 'async_trait>(
&'life0 self,
operation_id: OperationId
) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<'_, DepositState>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

fn get_withdraw_fee<'life0, 'async_trait>(
&'life0 self,
address: Address,
amount: Amount
) -> Pin<Box<dyn Future<Output = Result<PegOutFees>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Fetches the fees that would need to be paid to make the withdraw request using WalletClientExt::withdraw work right now.

Note that we do not receive a guarantee that these fees will be valid in the future, thus even the next second using these fees may fail. The caller should be prepared to retry with a new fee estimate.

fn withdraw<'life0, 'async_trait>(
&'life0 self,
address: Address,
amount: Amount,
fee: PegOutFees
) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Attempt to withdraw a given amount of Bitcoin to a destination address. The caller has to supply the fee rate to be used which can be fetched using WalletClientExt::get_withdraw_fee and should be acknowledged by the user since it can be unexpectedly high.

fn subscribe_withdraw_updates<'life0, 'async_trait>(
&'life0 self,
operation_id: OperationId
) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<'_, WithdrawState>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,