Crate fedimint_wallet_client

Modules

api

backup 🔒

cli 🔒

client_db

config

deposit 🔒

Legacy, state-machine based peg-ins, replaced by pegin_monitor but retained for time being to ensure existing peg-ins complete.

endpoint_constants

envs

keys

pegin_monitor 🔒

Peg-in monitor: a task monitoring deposit addresses for peg-ins.

tweakable

txoproof

withdraw 🔒

Structs

BitcoinTransactionData

PegOut

PegOutFees

PegOutSignatureItem

Rbf

Allows a user to bump the fees of a PendingTransaction

SpendableUTXO

TxOutputSummary

A transaction output, either unspent or consumed

UnknownWalletInputVariantError

UnknownWalletOutputOutcomeVariantError

UnknownWalletOutputVariantError

WalletClientContext

WalletClientInit

WalletClientModule

WalletClientModuleData

The non-resource, just plain-data parts of WalletClientModule

WalletCommonInit

WalletInputV0

WalletModuleTypes

WalletOperationMeta

WalletOutputOutcomeV0

Contains the Bitcoin transaction id of the transaction created by the withdraw request

WalletSummary

Summary of the coins within the wallet.

Enums

DepositStateV1

DepositStateV2

ProcessPegOutSigError

WalletClientStates

WalletConsensusItem

WalletCreationError

WalletInput

WalletInputError

WalletOperationMetaVariant

WalletOutput

WalletOutputError

WalletOutputOutcome

WalletOutputV0

WithdrawState

Constants

CONFIRMATION_TARGET

Used for estimating a feerate that will confirm within a target number of blocks.

DEPRECATED_RBF_ERROR

FEERATE_MULTIPLIER

To further mitigate the risk of a peg-out transaction getting stuck in the mempool, we multiply the feerate estimate returned from the backend by this value.

KIND

MODULE_CONSENSUS_VERSION

WALLET_TWEAK_CHILD_ID 🔒

Functions

check_address 🔒

get_next_peg_in_tweak_child_id 🔒

Returns the child index to derive the next peg-in tweak key from.

next_withdraw_state 🔒

proprietary_tweak_key

Type Aliases

PartialSig

PegInDescriptor