Crate fedimint_wallet_client

Modules

api

backup 🔒

cli 🔒

client_db

config

deposit 🔒

Legacy, state-machine based peg-ins, replaced by pegin_monitor but retained for time being to ensure existing peg-ins complete.

endpoint_constants

envs

keys

pegin_monitor 🔒

Peg-in monitor: a task monitoring deposit addresses for peg-ins.

tweakable

txoproof

withdraw 🔒

Structs

BitcoinTransactionData

PegOut

PegOutFees

PegOutSignatureItem

Rbf

Allows a user to bump the fees of a PendingTransaction

RecoverScanOutcome 🔒

SpendableUTXO

UnknownWalletInputVariantError

UnknownWalletOutputOutcomeVariantError

UnknownWalletOutputVariantError

WalletClientContext

WalletClientInit

WalletClientModule

WalletClientModuleData

The non-resource, just plain-data parts of WalletClientModule

WalletCommonInit

WalletInputV0

WalletModuleTypes

WalletOperationMeta

WalletOutputOutcomeV0

Contains the Bitcoin transaction id of the transaction created by the withdraw request

Enums

DepositState

ProcessPegOutSigError

WalletClientStates

WalletConsensusItem

WalletCreationError

WalletInput

WalletInputError

WalletOperationMetaVariant

WalletOutput

WalletOutputError

WalletOutputOutcome

WalletOutputV0

WithdrawState

Constants

CONFIRMATION_TARGET

Used for estimating a feerate that will confirm within a target number of blocks.

DEPRECATED_RBF_ERROR

FEERATE_MULTIPLIER

To further mitigate the risk of a peg-out transaction getting stuck in the mempool, we multiply the feerate estimate returned from the backend by this value.

KIND

MODULE_CONSENSUS_VERSION

RECOVER_MAX_GAP 🔒

We will check this many addresses after last actually used one before we give up

RECOVER_NUM_IDX_ADD_TO_LAST_USED 🔒

New client will start deriving new addresses from last used one plus that many idexes. This should be (much) less than MAX_GAP, but more than 0: We want to make sure we detect deposits that might have been made after multiple successive recoveries, but we want also to avoid accidental address re-use.

WALLET_TWEAK_CHILD_ID 🔒

Functions

check_address 🔒

get_next_peg_in_tweak_child_id 🔒

Returns the child index to derive the next peg-in tweak key from.

next_withdraw_state 🔒

proprietary_tweak_key

recover_scan_idxes_for_activity 🔒

A part of WalletClientInit::recover extracted out to be easy to test, as a side-effect free.

Type Aliases

PartialSig

PegInDescriptor