Crate fedimint_wallet_client

Modules

api

client_db

config

deposit 🔒

envs

keys

tweakable

txoproof

withdraw 🔒

Structs

BitcoinTransactionData

PegOut

PegOutFees

PegOutSignatureItem

Rbf

Allows a user to bump the fees of a PendingTransaction

SpendableUTXO

UnknownWalletInputVariantError

UnknownWalletOutputOutcomeVariantError

UnknownWalletOutputVariantError

WalletClientContext

WalletClientInit

WalletClientModule

WalletCommonInit

WalletInputV0

WalletModuleTypes

WalletOperationMeta

WalletOutputOutcomeV0

Contains the Bitcoin transaction id of the transaction created by the withdraw request

Enums

DepositState

ProcessPegOutSigError

WalletClientStates

WalletConsensusItem

WalletCreationError

WalletInput

WalletInputError

WalletOperationMetaVariant

WalletOutput

WalletOutputError

WalletOutputOutcome

WalletOutputV0

WithdrawState

Constants

CONFIRMATION_TARGET

KIND

MODULE_CONSENSUS_VERSION

WALLET_TWEAK_CHILD_ID 🔒

Functions

check_address 🔒

get_next_peg_in_tweak_child_id 🔒

Returns the child index to derive the next peg-in tweak key from.

next_deposit_state 🔒

next_withdraw_state 🔒

proprietary_tweak_key

Type Aliases

PartialSig

PegInDescriptor