Crate fedimint_wallet_client

Modules

api

config

db 🔒

deposit 🔒

keys

tweakable

txoproof

withdraw 🔒

Structs

PegOut

PegOutFees

PegOutSignatureItem

PendingTransaction

A peg-out tx that is ready to be broadcast with a tweak for the change UTXO

Rbf

Allows a user to bump the fees of a PendingTransaction

SpendableUTXO

UnsignedTransaction

A PSBT that is awaiting enough signatures from the federation to becoming a PendingTransaction

UnzipWalletConsensusItem

WalletClientContext

WalletClientGen

WalletClientModule

WalletCommonGen

WalletInput

WalletModuleTypes

WalletOutputOutcome

Contains the Bitcoin transaction id of the transaction created by the withdraw request

Enums

DepositState

ProcessPegOutSigError

WalletClientStates

WalletConsensusItem

WalletError

WalletOperationMeta

WalletOutput

WithdrawState

Constants

CONFIRMATION_TARGET

KIND

WALLET_TWEAK_CHILD_ID 🔒

Traits

IterUnzipWalletConsensusItem

WalletClientExt

Functions

check_address 🔒

get_next_peg_in_tweak_child_id 🔒

Returns the child index to derive the next peg-in tweak key from.

next_deposit_state 🔒

next_withdraw_state 🔒

proprietary_tweak_key

Type Definitions

PartialSig

PegInDescriptor