Crate fedimint_wallet_client

Modules

api

client_db 🔒

config

db

deposit 🔒

keys

tweakable

txoproof

withdraw 🔒

Structs

BitcoinTransactionData

PegOut

PegOutFees

PegOutSignatureItem

PendingTransaction

A peg-out tx that is ready to be broadcast with a tweak for the change UTXO

Rbf

Allows a user to bump the fees of a PendingTransaction

SpendableUTXO

UnknownWalletInputVariantError

UnknownWalletOutputOutcomeVariantError

UnknownWalletOutputVariantError

UnsignedTransaction

A PSBT that is awaiting enough signatures from the federation to becoming a PendingTransaction

WalletClientContext

WalletClientInit

WalletClientModule

WalletCommonInit

WalletInputV0

WalletModuleTypes

WalletOutputOutcomeV0

Contains the Bitcoin transaction id of the transaction created by the withdraw request

Enums

DepositState

ProcessPegOutSigError

WalletClientStates

WalletConsensusItem

WalletCreationError

WalletInput

WalletInputError

WalletOperationMeta

WalletOutput

WalletOutputError

WalletOutputOutcome

WalletOutputV0

WithdrawState

Constants

CONFIRMATION_TARGET

KIND

WALLET_TWEAK_CHILD_ID 🔒

Functions

check_address 🔒

get_next_peg_in_tweak_child_id 🔒

Returns the child index to derive the next peg-in tweak key from.

next_deposit_state 🔒

next_withdraw_state 🔒

proprietary_tweak_key

Type Definitions

PartialSig

PegInDescriptor