fedimint_wallet_client
Module client_db

Structs

ClaimedPegInData

ClaimedPegInKey

ClaimedPegInPrefix

DbKeyPrefixIter

An iterator over the variants of DbKeyPrefix

NextPegInTweakIndexKey

A counter tracking next index to use to derive a peg-in address

PegInTweakIndexData

PegInTweakIndexKey

Peg in index that was already allocated and is being tracked for deposits to claim

PegInTweakIndexPrefix

RecoveryFinalizedKey

RecoveryFinalizedKeyPrefix

RecoveryStateKey

RestoreStateKeyPrefix

TweakIdx

An index of a deposit address

Enums

DbKeyPrefix